requirements.txt
PyQt6>=6.4.0
PyOpenGL>=3.1.0
moderngl>=5.6.0
numpy>=1.21.0
numba>=0.56.0
scipy>=1.9.0
taichi>=1.4.0